Add test-client container and fix github CI action
