+++ .github/workflows/docker-build.yml
@@ -1,29 +1,34 @@
-name: Build and Push Docker Image
+name: Build and Push Docker Images to Github
+
+env:
+  VERSION: '1.0.0'
+  REPO_NAME: ${{ github.event.repository.name }}
+  MUJOCO_DOCKERFILE: "Dockerfile.simulation-mujoco"
+  PROJECT_OWNER: "openroboticmetaverse"
 
 on:
   push:
     branches:
-      - main  # Adjust this to match your main branch
+      - main
+  workflow_dispatch:
 
 jobs:
-  build:
+  push-mujoco-image:
     runs-on: ubuntu-latest
-
     steps:
-    - name: Checkout code
-      uses: actions/checkout@v2
+      - name: 'Checkout Repository'
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 0 
 
-    - name: Set up Docker Buildx
-      uses: docker/setup-buildx-action@v1
+      - name: 'Login to GitHub Container Registry'
+        uses: docker/login-action@v1
+        with:
+          registry: ghcr.io
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
 
-    - name: Build and push Docker image to GHCR
-      uses: docker/build-push-action@v2
-      with:
-        context: .
-        file: ./docker/Dockerfile
-        push: true
-        tags: ghcr.io/${{ github.repository_owner }}/mvp_mujoco_simulation:latest
-      env:
-        DOCKER_BUILDKIT: 1
-        DOCKER_USERNAME: ${{ github.repository_owner }}
-        DOCKER_PASSWORD: ${{ secrets.GITHUB_TOKEN }}
+      - name: 'Build and Push Image to Github Packages'
+        run: |
+          docker build -f ./docker/${{ env.MUJOCO_DOCKERFILE }} -t ghcr.io/${{ env.PROJECT_OWNER }}/${{ env.REPO_NAME }}:${{ env.VERSION }} .
+          docker push ghcr.io/${{ env.PROJECT_OWNER }}/${{ env.REPO_NAME }}:latest
