+++ docker/Dockerfile.simulation-mujoco
@@ -13,6 +13,7 @@ COPY ./config /sim_ws/config
 
 WORKDIR /sim_ws
 
+RUN pip install --upgrade pip
 RUN pip install -r requirements.txt
-ENTRYPOINT ["python3", "src/mujoco_simulation.py"]
 
+ENTRYPOINT ["python3", "src/mujoco_simulation.py"]
