+++ docker/Dockerfile.test-client
@@ -1,0 +1,7 @@
+FROM python:3.12-slim
+
+WORKDIR /app
+COPY ./test/test_client.py .
+RUN pip install websockets
+
+CMD ["python3", "test_client.py"]
